# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sky Wings API",
        "description": "API documentation for Sky Wings application",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Get all admins",
                "operationId": "4d591b8bc6e9bc9494d2dd87c1032f02",
                "responses": {
                    "200": {
                        "description": "List of admins"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admins"
                ],
                "summary": "Create a new admin",
                "operationId": "8405ee30d2a8cacc8f0356194fb256ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "admin123"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin successfully created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/admin/{id}": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Get a specific admin by ID",
                "operationId": "25771fd3394a046c032b0722a00cf908",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin details"
                    },
                    "404": {
                        "description": "Admin not found"
                    }
                }
            }
        },
        "/api/admin/signIn": {
            "post": {
                "tags": [
                    "Admins"
                ],
                "summary": "Admin login",
                "operationId": "a849ba553a87c18bb062b294de5c5dd9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin logged in successfully"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
...
        "/api/auth/signIn": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate a user by username or email",
                "operationId": "b38f09fdcdaf6142142287783648fb14",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "description": "Username or email",
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User successfully authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User logged in successfully."
                                        },
                                        "data": {
                                            "type": "integer",
                                            "example": 42
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email/Username or password is invalid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}